moving current carousel to old and replacing with the new code
