--- aemedge/blocks/carousel-new/carousel-new.css
@@ -1,57 +1,0 @@
-/* Carousel */
-
-.carousel-slides {
-  display: flex;
-  gap: 8px;
-  padding: 16px;
-  list-style: none;
-  overflow-x: scroll;
-  scroll-snap-type: x mandatory;
-  
-  /* Hide scrollbar in Firefox */
-  scrollbar-width: none;
-  
-  /* Hide scrollbar in IE and Edge */
-  -ms-overflow-style: none;
-}
-
-.carousel-slides::-webkit-scrollbar {
-  display: none;
-}
-
-.carousel-slide {
-  flex-shrink: 0;
-  width: 33.34%;
-  height: 30%;
-  background-color: #FFF;
-  scroll-snap-align: center;
-}
-
-.carousel-slide > div {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-}
-
-.carousel-new .button {
-  position: absolute;
-  top: 50%;
-  width: 3rem;
-  height: 3rem;
-  transform: translateY(-50%);
-  color: black;
-  background-color: unset;
-  margin: 0;
-  padding: 0;
-  z-index: 1;
-}
-
-.carousel-new .button-previous {
-  left: 1.5rem;
-  transform: rotate(180deg);
-}
-
-.carousel-new .button-next {
-  right: 1.5rem;
-}
