+++ aemedge/blocks/carousel/carousel.css
@@ -1,182 +1,57 @@
-.carousel {
-  padding: 50px 10% 0 6%;
-}
+/* Carousel */
 
-.carousel .carousel-slides-container {
-  position: relative;
-}
-
-.carousel .carousel-slides,
-.carousel .carousel-slide-indicators {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.carousel .carousel-slides {
+.carousel-slides {
   display: flex;
-  scroll-behavior: smooth;
+  gap: 8px;
+  padding: 16px;
+  list-style: none;
+  overflow-x: scroll;
   scroll-snap-type: x mandatory;
-  overflow: scroll clip;
+  
+  /* Hide scrollbar in Firefox */
+  scrollbar-width: none;
+  
+  /* Hide scrollbar in IE and Edge */
+  -ms-overflow-style: none;
 }
 
-.carousel .carousel-slides::-webkit-scrollbar {
+.carousel-slides::-webkit-scrollbar {
   display: none;
 }
 
-.carousel .carousel-slide {
-  scroll-snap-align: start;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  position: relative;
-  min-width: 50%;
-
-  /* align-items: flex-start; */
-
-  /* min-height: min(40rem, calc(100svh - var(--nav-height))); */
+.carousel-slide {
+  flex-shrink: 0;
+  width: 33.34%;
+  height: 30%;
+  background-color: #FFF;
+  scroll-snap-align: center;
 }
 
-.carousel .carousel-slide:has(.carousel-slide-content[data-align="center"]) {
+.carousel-slide > div {
+  display: flex;
   align-items: center;
-}
-
-.carousel .carousel-slide:has(.carousel-slide-content[data-align="right"]) {
-  align-items: flex-end;
-}
-
-.carousel .carousel-slide .carousel-slide-image {
-  display: flex;
-  aspect-ratio: 2 / 3;
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  text-decoration: none;
-  pointer-events: none;
-  position: relative;
-  text-align: center;
-  color: white;
-}
-
-.carousel .carousel-slide .carousel-slide-image picture {
-  /* position: absolute;
-  inset: 0; */
-  pointer-events: none;
-  min-width: 100%;
-  width: 100%;
-  min-height: 100%;
+  justify-content: center;
   height: 100%;
 }
 
-.carousel .carousel-slide .carousel-slide-image picture > img {
-  /* height: 100%;
-  width: 100%;
-  object-fit: cover; */
-  pointer-events: auto;
-  min-height: calc(100% - 1.5rem);
-  height: calc(100% - 1.5rem);
-  min-width: calc(100% - 1.5rem);
-  width: calc(100% - 1.5rem);
-  margin: 0.75rem;
-  border-radius: 0.25em;
-  object-fit: cover;
-  background: rgb(239 239 241);
-  filter: drop-shadow(0 4px 4px rgb(0 0 0 / 20%));
-}
-
-.carousel .carousel-slide .carousel-slide-content {
-  z-index: 1;
-  padding: 1rem;
-  margin: 1.5rem 3rem;
-  color: white;
-  background-color: rgba(0 0 0 / 50%);
-  position: relative;
-  width: var(--slide-content-width, auto);
-}
-
-.carousel .carousel-slide-indicators {
-  display: none;
-}
-
-.carousel .carousel-slide-indicator span,
-.carousel .carousel-navigation-buttons span {
-  border: 0;
-  clip: rect(0 0 0 0);
-  clip-path: inset(50%);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  width: 1px;
-  white-space: nowrap;
-}
-
-.carousel .carousel-navigation-buttons {
+.carousel-new .button {
   position: absolute;
   top: 50%;
+  width: 3rem;
+  height: 3rem;
   transform: translateY(-50%);
-  left: -10px;
-  right: -10px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
+  color: black;
+  background-color: unset;
+  margin: 0;
+  padding: 0;
   z-index: 1;
 }
 
-/* stylelint-disable-next-line no-descending-specificity */
-.carousel .carousel-navigation-buttons button {
-  backface-visibility: hidden;
-  top: calc(50% - 1.25rem);
-  left: 0.625rem;
-  min-width: 2.75rem;
-  width: 2.75rem;
-  min-height: 2.75rem;
-  height: 2.75rem;
-  line-height: 0;
-  border: none;
-  border-radius: 50%;
-  background: rgb(245 245 246 / 80%);
-  cursor: pointer;
-  filter: drop-shadow(0 2px 10px rgb(0 0 0 / 30%));
-  backdrop-filter: blur(5px);
+.carousel-new .button-previous {
+  left: 1.5rem;
+  transform: rotate(180deg);
 }
 
-.carousel .carousel-navigation-buttons button::after {
-  display: block;
-  content: "";
-  border: 2.5px black solid;
-  border-bottom: 0;
-  border-left: 0;
-  height: 0.5rem;
-  width: 0.5rem;
-  position: absolute;
-  top: 50%;
-  left: calc(50% + 2px);
-  transform: translate(-50%, -50%) rotate(-135deg);
-}
-
-.carousel .carousel-navigation-buttons button.slide-next::after {
-  transform: translate(-50%, -50%)  rotate(45deg);
-  left: calc(50% - 3px);
-}
-
-@media (width >= 768px) {
-  .carousel {
-    padding: 40px 40px 0 30px;
-  }
-
-  .carousel .carousel-slide {
-    min-width: 33.334%;
-  }
-
-  .carousel .carousel-slide .carousel-slide-content {
-    --slide-content-width: 50%;
-
-    margin: 2.5rem 5rem;
-  }
-
-  .carousel .carousel-slide .carousel-slide-content[data-align="justify"] {
-    --slide-content-width: auto;
-  }
+.carousel-new .button-next {
+  right: 1.5rem;
 }
